move docket scopes modules into the DistributionScopes class
